Defined each sequelize model needed
