+++ db/Product.js
@@ -1,30 +1,5 @@
 const {DataTypes} = require('sequelize');
-const sequelize = require('./db.js');
+const {sequelize, Product} = require('./db.js');
 
-const Product = sequelize.define('Product', {
-  id: {
-    type: DataTypes.INTEGER,
-    allowNull: false,
-    primaryKey: true,
-    validate: {
-      isInt: true
-    }
-  },
-  name: {
-    type: DataTypes.STRING
-  },
-  slogan: {
-    type: DataTypes.STRING
-  },
-  description: {
-    type: DataTypes.STRING(1000)
-  },
-  category: {
-    type: DataTypes.STRING
-  },
-  default_price: {
-    type: DataTypes.INTEGER
-  }
-})
 
-module.exports = Product;
+
